change float to double

+++ APITest/DataDTO/MyDataDTO.cs
@@ -4,7 +4,7 @@
     {
         public string Title { get; set; }
 
-        public float Length { get; set; }
+        public double Length { get; set; }
 
         public bool IsTrue { get; set; }
     }
